ran with flags

+++ src/input_validation_and_map_creation/check_if_map_index_exists.c
@@ -6,7 +6,7 @@
 /*   By: kmilchev <kmilchev@student.42wolfsburg.de> +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/07 20:13:33 by kmilchev          #+#    #+#             */
-/*   Updated: 2022/07/07 22:20:42 by kmilchev         ###   ########.fr       */
+/*   Updated: 2022/07/14 18:17:42 by kmilchev         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 bool	square_exists(t_mlx *mlx_info, int i, int j)
 {
 	if ((i > 0 && i < mlx_info->map_height - 1)
-		&& (j > 0 && j < ft_strlen(mlx_info->map_s[i])))
+		&& (j > 0 && j < (int)ft_strlen(mlx_info->map_s[i])))
 		return (true);
 	return (false);
 }
@@ -24,7 +24,7 @@ bool	square_exists(t_mlx *mlx_info, int i, int j)
 bool	square_below_exists(t_mlx *mlx_info, int i, int j)
 {
 	if ((i + 1 > 0 && i + 1 <= mlx_info->map_height - 1)
-		&& (j > 0 && j < ft_strlen(mlx_info->map_s[i + 1]) - 1))
+		&& (j > 0 && j < (int)ft_strlen(mlx_info->map_s[i + 1]) - 1))
 		return (true);
 	return (false);
 }
@@ -32,7 +32,7 @@ bool	square_below_exists(t_mlx *mlx_info, int i, int j)
 bool	square_above_exists(t_mlx *mlx_info, int i, int j)
 {
 	if ((i - 1 >= 0 && i - 1 <= mlx_info->map_height - 1)
-		&& (j < ft_strlen(mlx_info->map_s[i - 1])))
+		&& (j < (int)ft_strlen(mlx_info->map_s[i - 1])))
 		return (true);
 	return (false);
 }
@@ -40,7 +40,7 @@ bool	square_above_exists(t_mlx *mlx_info, int i, int j)
 bool	square_left_exists(t_mlx *mlx_info, int i, int j)
 {
 	if ((i > 0 && i < mlx_info->map_height - 1)
-		&& (j - 1 >= 0 && j - 1 <= ft_strlen(mlx_info->map_s[i - 1]) - 1))
+		&& (j - 1 >= 0 && j - 1 <= (int)ft_strlen(mlx_info->map_s[i - 1]) - 1))
 		return (true);
 	return (false);
 }
@@ -48,7 +48,7 @@ bool	square_left_exists(t_mlx *mlx_info, int i, int j)
 bool	square_right_exists(t_mlx *mlx_info, int i, int j)
 {
 	if ((i > 0 && i < mlx_info->map_height - 1)
-		&& (j + 1 > 0 && j + 1 < ft_strlen(mlx_info->map_s[i - 1])))
+		&& (j + 1 > 0 && j + 1 < (int)ft_strlen(mlx_info->map_s[i - 1])))
 		return (true);
 	return (false);
 }
